upd: небольшой рефакторинг, добился передачу данных до 1000000

+++ client_server/client/mainwindow.h
@@ -10,6 +10,7 @@
 #include <QJsonDocument>
 #include <QJsonObject>
 #include <QJsonArray>
+#include <QScopedPointer>
 
 namespace Ui {
 class MainWindow;
@@ -19,6 +20,11 @@ class MainWindow : public QMainWindow
 {
     Q_OBJECT
 
+    /*!
+     * Перечисление кодов алгоритмов
+     * TODO: По уму, чтобы дублирования кода не было, нужно
+     * вынести в отдельный файл, доступный и  для клиента и для сервера.
+     */
     enum {
         NO_ALGO = 0,
         SORT_CHARS_BY_DESCENDING,
@@ -34,12 +40,42 @@ public:
 private:
     Ui::MainWindow *ui;
 
+    /*!
+     * \brief Указатель на объект сокета
+     */
     QTcpSocket         *m_tcpSocket;
-    quint16             m_blockSize = 0;
+    /*!
+     * \brief Инициализация размера блока данных
+     * CONFIG += c++11
+     */
+    quint32             m_blockSize = 0;
+
+    /*!
+     * \brief хеш алгоритмов
+     * TODO: По уму, чтобы дублирования кода не было, нужно
+     * вынести в отдельный файл, доступный и  для клиента и для сервера.
+     */
     QHash<int,QString>  m_algorithms;
 
+
+    /*!
+     * \brief Создание Json пакета для сервера
+     * \param[in] algo - алгоритм
+     * \param[in] text - текст
+     * \return
+     * TODO: По уму, чтобы дублирования кода не было, нужно
+     * вынести в отдельный файл, доступный и  для клиента и для сервера.
+     */
     QByteArray createJsonPacket(QString algo, QString text);
 
+    /*!
+     * \brief Разбор Json пакета от сервера
+     * \param[in] bytes - данные
+     * \param[in] algo - алгоритм
+     * \param[in] text - текст данных
+     * TODO: По уму, чтобы дублирования кода не было, нужно
+     * вынести в отдельный файл, доступный и  для клиента и для сервера.
+     */
     void parseJsonPacket(QByteArray &bytes, QString &algo, QStringList &text);
 
 private slots:
@@ -63,8 +99,15 @@ private slots:
      */
     void connected();
 
-    /* Обработчики нажатия кнопок */
+    /*!
+     * \brief Нажатие на кнопку соединения
+     */
     void on_connectButton_clicked();
+
+    /*!
+     * \brief Нажатие на кнопку отправки данных
+     * TODO: Нужно сделать проверки на то что соединение установлено
+     */
     void on_sendButton_clicked();
 };
